+++ NEWS.md
@@ -1,6 +1,8 @@
 # Current
 
 - An advanced plotting tutorial
+- Try harder to kill maxima on closing it
+- Try harder to clean up maxima's temporary files on closing it
 - Resolved a CMake error on configuration with -B switch (#1917)
 - More menu icons
 - A better dialogue if Maxima isn't found